infra/resources/training/labelling_complete/requirements.txt
boto3
loguru
albumentations
opencv-python
jsonlines
Pillow